backend/requirements.txt
asgiref==3.3.1
attrs==21.2.0
autopep8==1.5.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
django_elasticsearch_dsl_drf===0.22.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
Django==3.1.7
django-cors-headers==3.7.0
django-elasticsearch-dsl==7.2.0
django-elasticsearch-dsl-drf==0.22.2
django-nine==0.2.5
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.12.2
djangorestframework-jsonapi==4.1.0
djangorestframework-simplejwt==4.6.0
elasticsearch==7.14.0
elasticsearch-dsl==7.4.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.2.1
MarkupSafe==2.0.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==21.3
psycopg2==2.8.6
pycodestyle==2.7.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
ruamel.yaml==0.17.17
simplejson==3.17.6
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
pandas==1.3.5
drf_spectacular==0.21.1
django-multiselectfield==0.1.12
python-dotenv==0.19.2